Add pyi files for Athena read
