+++ awswrangler/athena/_read.py
@@ -6,7 +6,7 @@ import logging
 import sys
 import uuid
 from datetime import date
-from typing import Any, Dict, Iterator, List, Literal, Optional, Union, cast, overload
+from typing import Any, Dict, Iterator, List, Optional, Union, cast
 
 import boto3
 import botocore.exceptions
@@ -592,61 +592,6 @@ def _unload(
     return query_metadata
 
 
-@overload
-def get_query_results(
-    query_execution_id: str,
-    use_threads: Union[bool, int] = ...,
-    boto3_session: Optional[boto3.Session] = ...,
-    categories: Optional[List[str]] = ...,
-    chunksize: Union[None, Literal[False]] = ...,
-    s3_additional_kwargs: Optional[Dict[str, Any]] = ...,
-    pyarrow_additional_kwargs: Optional[Dict[str, Any]] = ...,
-) -> pd.DataFrame:
-    ...
-
-
-@overload
-def get_query_results(
-    query_execution_id: str,
-    *,
-    use_threads: Union[bool, int] = ...,
-    boto3_session: Optional[boto3.Session] = ...,
-    categories: Optional[List[str]] = ...,
-    chunksize: Literal[True],
-    s3_additional_kwargs: Optional[Dict[str, Any]] = ...,
-    pyarrow_additional_kwargs: Optional[Dict[str, Any]] = ...,
-) -> Iterator[pd.DataFrame]:
-    ...
-
-
-@overload
-def get_query_results(
-    query_execution_id: str,
-    *,
-    use_threads: Union[bool, int] = ...,
-    boto3_session: Optional[boto3.Session] = ...,
-    categories: Optional[List[str]] = ...,
-    chunksize: bool,
-    s3_additional_kwargs: Optional[Dict[str, Any]] = ...,
-    pyarrow_additional_kwargs: Optional[Dict[str, Any]] = ...,
-) -> Union[pd.DataFrame, Iterator[pd.DataFrame]]:
-    ...
-
-
-@overload
-def get_query_results(
-    query_execution_id: str,
-    *,
-    use_threads: Union[bool, int] = ...,
-    boto3_session: Optional[boto3.Session] = ...,
-    categories: Optional[List[str]] = ...,
-    chunksize: int,
-    s3_additional_kwargs: Optional[Dict[str, Any]] = ...,
-    pyarrow_additional_kwargs: Optional[Dict[str, Any]] = ...,
-) -> Iterator[pd.DataFrame]:
-    ...
-
-
 @apply_configs
 @_utils.validate_distributed_kwargs(
     unsupported_kwargs=["boto3_session"],
@@ -738,145 +683,6 @@ def get_query_results(
             s3_additional_kwargs=s3_additional_kwargs,
         )
     raise exceptions.UndetectedType(f"""Unable to get results for: {query_info["Query"]}.""")
-
-
-@overload
-def read_sql_query(  # pylint: disable=too-many-arguments
-    sql: str,
-    database: str,
-    ctas_approach: bool = ...,
-    unload_approach: bool = ...,
-    ctas_parameters: Optional[typing.AthenaCTASSettings] = ...,
-    unload_parameters: Optional[typing.AthenaUNLOADSettings] = ...,
-    categories: Optional[List[str]] = ...,
-    chunksize: Union[None, Literal[False]] = ...,
-    s3_output: Optional[str] = ...,
-    workgroup: Optional[str] = ...,
-    encryption: Optional[str] = ...,
-    kms_key: Optional[str] = ...,
-    keep_files: bool = ...,
-    use_threads: Union[bool, int] = ...,
-    boto3_session: Optional[boto3.Session] = ...,
-    athena_cache_settings: Optional[typing.AthenaCacheSettings] = ...,
-    data_source: Optional[str] = ...,
-    athena_query_wait_polling_delay: float = ...,
-    params: Optional[Dict[str, Any]] = ...,
-    s3_additional_kwargs: Optional[Dict[str, Any]] = ...,
-    pyarrow_additional_kwargs: Optional[Dict[str, Any]] = ...,
-) -> pd.DataFrame:
-    ...
-
-
-@overload
-def read_sql_query(
-    sql: str,
-    database: str,
-    *,
-    ctas_approach: bool = ...,
-    unload_approach: bool = ...,
-    ctas_parameters: Optional[typing.AthenaCTASSettings] = ...,
-    unload_parameters: Optional[typing.AthenaUNLOADSettings] = ...,
-    categories: Optional[List[str]] = ...,
-    chunksize: Literal[True],
-    s3_output: Optional[str] = ...,
-    workgroup: Optional[str] = ...,
-    encryption: Optional[str] = ...,
-    kms_key: Optional[str] = ...,
-    keep_files: bool = ...,
-    use_threads: Union[bool, int] = ...,
-    boto3_session: Optional[boto3.Session] = ...,
-    athena_cache_settings: Optional[typing.AthenaCacheSettings] = ...,
-    data_source: Optional[str] = ...,
-    athena_query_wait_polling_delay: float = ...,
-    params: Optional[Dict[str, Any]] = ...,
-    s3_additional_kwargs: Optional[Dict[str, Any]] = ...,
-    pyarrow_additional_kwargs: Optional[Dict[str, Any]] = ...,
-) -> Iterator[pd.DataFrame]:
-    ...
-
-
-@overload
-def read_sql_query(
-    sql: str,
-    database: str,
-    *,
-    ctas_approach: bool = ...,
-    unload_approach: bool = ...,
-    ctas_parameters: Optional[typing.AthenaCTASSettings] = ...,
-    unload_parameters: Optional[typing.AthenaUNLOADSettings] = ...,
-    categories: Optional[List[str]] = ...,
-    chunksize: bool,
-    s3_output: Optional[str] = ...,
-    workgroup: Optional[str] = ...,
-    encryption: Optional[str] = ...,
-    kms_key: Optional[str] = ...,
-    keep_files: bool = ...,
-    use_threads: Union[bool, int] = ...,
-    boto3_session: Optional[boto3.Session] = ...,
-    athena_cache_settings: Optional[typing.AthenaCacheSettings] = ...,
-    data_source: Optional[str] = ...,
-    athena_query_wait_polling_delay: float = ...,
-    params: Optional[Dict[str, Any]] = ...,
-    s3_additional_kwargs: Optional[Dict[str, Any]] = ...,
-    pyarrow_additional_kwargs: Optional[Dict[str, Any]] = ...,
-) -> Union[pd.DataFrame, Iterator[pd.DataFrame]]:
-    ...
-
-
-@overload
-def read_sql_query(
-    sql: str,
-    database: str,
-    *,
-    ctas_approach: bool = ...,
-    unload_approach: bool = ...,
-    ctas_parameters: Optional[typing.AthenaCTASSettings] = ...,
-    unload_parameters: Optional[typing.AthenaUNLOADSettings] = ...,
-    categories: Optional[List[str]] = ...,
-    chunksize: int,
-    s3_output: Optional[str] = ...,
-    workgroup: Optional[str] = ...,
-    encryption: Optional[str] = ...,
-    kms_key: Optional[str] = ...,
-    keep_files: bool = ...,
-    use_threads: Union[bool, int] = ...,
-    boto3_session: Optional[boto3.Session] = ...,
-    athena_cache_settings: Optional[typing.AthenaCacheSettings] = ...,
-    data_source: Optional[str] = ...,
-    athena_query_wait_polling_delay: float = ...,
-    params: Optional[Dict[str, Any]] = ...,
-    s3_additional_kwargs: Optional[Dict[str, Any]] = ...,
-    pyarrow_additional_kwargs: Optional[Dict[str, Any]] = ...,
-) -> Iterator[pd.DataFrame]:
-    ...
-
-
-@overload
-def read_sql_query(
-    sql: str,
-    database: str,
-    *,
-    ctas_approach: bool = ...,
-    unload_approach: bool = ...,
-    ctas_parameters: Optional[typing.AthenaCTASSettings] = ...,
-    unload_parameters: Optional[typing.AthenaUNLOADSettings] = ...,
-    categories: Optional[List[str]] = ...,
-    chunksize: Optional[Union[int, bool]],
-    s3_output: Optional[str] = ...,
-    workgroup: Optional[str] = ...,
-    encryption: Optional[str] = ...,
-    kms_key: Optional[str] = ...,
-    keep_files: bool = ...,
-    use_threads: Union[bool, int] = ...,
-    boto3_session: Optional[boto3.Session] = ...,
-    athena_cache_settings: Optional[typing.AthenaCacheSettings] = ...,
-    data_source: Optional[str] = ...,
-    athena_query_wait_polling_delay: float = ...,
-    params: Optional[Dict[str, Any]] = ...,
-    s3_additional_kwargs: Optional[Dict[str, Any]] = ...,
-    pyarrow_additional_kwargs: Optional[Dict[str, Any]] = ...,
-) -> Union[pd.DataFrame, Iterator[pd.DataFrame]]:
-    ...
 
 
 @apply_configs
@@ -1192,136 +998,6 @@ def read_sql_query(  # pylint: disable=too-many-arguments,too-many-locals
         boto3_session=boto3_session,
         pyarrow_additional_kwargs=pyarrow_additional_kwargs,
     )
-
-
-@overload
-def read_sql_table(
-    table: str,
-    database: str,
-    *,
-    unload_approach: bool = ...,
-    unload_parameters: Optional[typing.AthenaUNLOADSettings] = ...,
-    ctas_approach: bool = ...,
-    ctas_parameters: Optional[typing.AthenaCTASSettings] = ...,
-    categories: Optional[List[str]] = ...,
-    chunksize: Union[None, Literal[False]] = ...,
-    s3_output: Optional[str] = ...,
-    workgroup: Optional[str] = ...,
-    encryption: Optional[str] = ...,
-    kms_key: Optional[str] = ...,
-    keep_files: bool = ...,
-    use_threads: Union[bool, int] = ...,
-    boto3_session: Optional[boto3.Session] = ...,
-    athena_cache_settings: Optional[typing.AthenaCacheSettings] = ...,
-    data_source: Optional[str] = ...,
-    s3_additional_kwargs: Optional[Dict[str, Any]] = ...,
-    pyarrow_additional_kwargs: Optional[Dict[str, Any]] = ...,
-) -> pd.DataFrame:
-    ...
-
-
-@overload
-def read_sql_table(
-    table: str,
-    database: str,
-    *,
-    unload_approach: bool = ...,
-    unload_parameters: Optional[typing.AthenaUNLOADSettings] = ...,
-    ctas_approach: bool = ...,
-    ctas_parameters: Optional[typing.AthenaCTASSettings] = ...,
-    categories: Optional[List[str]] = ...,
-    chunksize: Literal[True],
-    s3_output: Optional[str] = ...,
-    workgroup: Optional[str] = ...,
-    encryption: Optional[str] = ...,
-    kms_key: Optional[str] = ...,
-    keep_files: bool = ...,
-    use_threads: Union[bool, int] = ...,
-    boto3_session: Optional[boto3.Session] = ...,
-    athena_cache_settings: Optional[typing.AthenaCacheSettings] = ...,
-    data_source: Optional[str] = ...,
-    s3_additional_kwargs: Optional[Dict[str, Any]] = ...,
-    pyarrow_additional_kwargs: Optional[Dict[str, Any]] = ...,
-) -> Iterator[pd.DataFrame]:
-    ...
-
-
-@overload
-def read_sql_table(
-    table: str,
-    database: str,
-    *,
-    unload_approach: bool = ...,
-    unload_parameters: Optional[typing.AthenaUNLOADSettings] = ...,
-    ctas_approach: bool = ...,
-    ctas_parameters: Optional[typing.AthenaCTASSettings] = ...,
-    categories: Optional[List[str]] = ...,
-    chunksize: bool,
-    s3_output: Optional[str] = ...,
-    workgroup: Optional[str] = ...,
-    encryption: Optional[str] = ...,
-    kms_key: Optional[str] = ...,
-    keep_files: bool = ...,
-    use_threads: Union[bool, int] = ...,
-    boto3_session: Optional[boto3.Session] = ...,
-    athena_cache_settings: Optional[typing.AthenaCacheSettings] = ...,
-    data_source: Optional[str] = ...,
-    s3_additional_kwargs: Optional[Dict[str, Any]] = ...,
-    pyarrow_additional_kwargs: Optional[Dict[str, Any]] = ...,
-) -> Union[pd.DataFrame, Iterator[pd.DataFrame]]:
-    ...
-
-
-@overload
-def read_sql_table(
-    table: str,
-    database: str,
-    *,
-    unload_approach: bool = ...,
-    unload_parameters: Optional[typing.AthenaUNLOADSettings] = ...,
-    ctas_approach: bool = ...,
-    ctas_parameters: Optional[typing.AthenaCTASSettings] = ...,
-    categories: Optional[List[str]] = ...,
-    chunksize: int,
-    s3_output: Optional[str] = ...,
-    workgroup: Optional[str] = ...,
-    encryption: Optional[str] = ...,
-    kms_key: Optional[str] = ...,
-    keep_files: bool = ...,
-    use_threads: Union[bool, int] = ...,
-    boto3_session: Optional[boto3.Session] = ...,
-    athena_cache_settings: Optional[typing.AthenaCacheSettings] = ...,
-    data_source: Optional[str] = ...,
-    s3_additional_kwargs: Optional[Dict[str, Any]] = ...,
-    pyarrow_additional_kwargs: Optional[Dict[str, Any]] = ...,
-) -> Iterator[pd.DataFrame]:
-    ...
-
-
-@overload
-def read_sql_table(
-    table: str,
-    database: str,
-    *,
-    unload_approach: bool = ...,
-    unload_parameters: Optional[typing.AthenaUNLOADSettings] = ...,
-    ctas_approach: bool = ...,
-    ctas_parameters: Optional[typing.AthenaCTASSettings] = ...,
-    categories: Optional[List[str]] = ...,
-    chunksize: Optional[Union[int, bool]],
-    s3_output: Optional[str] = ...,
-    workgroup: Optional[str] = ...,
-    encryption: Optional[str] = ...,
-    kms_key: Optional[str] = ...,
-    keep_files: bool = ...,
-    use_threads: Union[bool, int] = ...,
-    boto3_session: Optional[boto3.Session] = ...,
-    athena_cache_settings: Optional[typing.AthenaCacheSettings] = ...,
-    data_source: Optional[str] = ...,
-    s3_additional_kwargs: Optional[Dict[str, Any]] = ...,
-    pyarrow_additional_kwargs: Optional[Dict[str, Any]] = ...,
-) -> Union[pd.DataFrame, Iterator[pd.DataFrame]]:
-    ...
 
 
 @apply_configs
